refactor: jpaRepository infra 영역으로 옮김
