+++ src/main/java/kitchenpos/domain/OrderRepository.java
@@ -15,7 +15,3 @@ public interface OrderRepository {
 
     List<Order> findAll();
 }
-
-interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
-    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
-}
