+++ src/main/java/kitchenpos/infra/JpaMenuGroupRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.infra;
+
+import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.MenuGroupRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, UUID> {
+}
