+++ src/main/java/kitchenpos/infra/JpaOrderRepository.java
@@ -1,0 +1,13 @@
+package kitchenpos.infra;
+
+import kitchenpos.domain.Order;
+import kitchenpos.domain.OrderRepository;
+import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.OrderTable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
+}
