+++ src/main/java/kitchenpos/infra/JpaOrderTableRepository.java
@@ -1,0 +1,11 @@
+package kitchenpos.infra;
+
+import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.OrderTableRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
+
+}
