+++ src/main/java/kitchenpos/infra/JpaProductRepository.java
@@ -1,0 +1,12 @@
+package kitchenpos.infra;
+
+import kitchenpos.domain.Product;
+import kitchenpos.domain.ProductRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
+    List<Product> findAllByIdIn(List<UUID> ids);
+}
